[AutoDiff] draft of lifting samefile derivative constriant
