+++ include/swift/AST/DiagnosticsSema.def
@@ -3016,12 +3016,15 @@ NOTE(derivative_attr_result_func_type_mismatch_note,none,
      "%0 does not have expected type %1", (Identifier, Type))
 NOTE(derivative_attr_result_func_original_note,none,
      "%0 defined here", (DeclName))
-ERROR(derivative_attr_not_in_same_file_as_original,none,
-      "derivative not in the same file as the original function", ())
 ERROR(derivative_attr_original_stored_property_unsupported,none,
       "cannot register derivative for stored property %0", (DeclName))
 ERROR(derivative_attr_original_already_has_derivative,none,
       "a derivative already exists for %0", (DeclName))
+ERROR(derivative_attr_visibility_too_broad,none,
+      "derivative function visibility must be at least as restrictive as original function "
+      "visibility", ())
+NOTE(derivative_attr_visibility_too_broad_note,none,
+     "original function defined here", ())
 
 // @transpose
 ERROR(transpose_params_clause_param_not_differentiable,none,
