+++ test/AutoDiff/Inputs/derivative_attr_type_checking/module2/module2.swift
@@ -1,0 +1,6 @@
+import module1
+
+@derivative(of: functionDefinedInModule1_publicDerivativeInModule2)
+public func dfModule2(_ x: Float) -> (value: Float, pullback: (Float) -> Float) {
+  fatalError()
+}
